📦 hoist site constants out of seo module
