+++ app/layout.tsx
@@ -16,11 +16,11 @@ import Script from "next/script";
 import "./globals.css";
 import {
   DEFAULT_DESCRIPTION,
-  DEFAULT_OG_IMAGE,
-  OG_IMAGE_SIZE,
   SITE_NAME,
+  SITE_TAGLINE,
   SITE_URL,
-} from "@/lib/seo/constants";
+} from "@/lib/constants";
+import { DEFAULT_OG_IMAGE, OG_IMAGE_SIZE } from "@/lib/seo/constants";
 
 const manrope = Manrope({
   subsets: ["latin"],
@@ -35,7 +35,7 @@ if (process.env.NODE_ENV !== "production") {
 
 export const metadata: Metadata = {
   metadataBase: new URL(`${SITE_URL}/`),
-  title: `${SITE_NAME} | Watch Movies and TV Shows`,
+  title: `${SITE_NAME} | ${SITE_TAGLINE}`,
   description: DEFAULT_DESCRIPTION,
   icons: {
     icon: [
@@ -60,7 +60,7 @@ export const metadata: Metadata = {
     type: "website",
     url: SITE_URL,
     siteName: SITE_NAME,
-    title: `${SITE_NAME} | Watch Movies and TV Shows`,
+    title: `${SITE_NAME} | ${SITE_TAGLINE}`,
     description: DEFAULT_DESCRIPTION,
     images: [
       {
@@ -68,13 +68,13 @@ export const metadata: Metadata = {
         width: OG_IMAGE_SIZE.width,
         height: OG_IMAGE_SIZE.height,
         type: "image/webp",
-        alt: `${SITE_NAME} | Watch Movies and TV Shows`,
+        alt: `${SITE_NAME} | ${SITE_TAGLINE}`,
       },
     ],
   },
   twitter: {
     card: "summary_large_image",
-    title: `${SITE_NAME} | Watch Movies and TV Shows`,
+    title: `${SITE_NAME} | ${SITE_TAGLINE}`,
     description: DEFAULT_DESCRIPTION,
     images: [DEFAULT_OG_IMAGE],
   },
